Move GraphQL object to gql_models
