+++ src/ai/backend/manager/models/gql_models/__init__.py
@@ -1,0 +1,28 @@
+from . import domain as _domain
+from . import group as _group
+from . import image as _image
+from . import kernel as _kernel
+from . import keypair as _keypair
+from . import scaling_group as _scaling_group
+from . import session as _session
+from . import user as _user
+
+__all__ = (
+    *_domain.__all__,
+    *_group.__all__,
+    *_image.__all__,
+    *_kernel.__all__,
+    *_keypair.__all__,
+    *_scaling_group.__all__,
+    *_session.__all__,
+    *_user.__all__,
+)
+
+from .domain import *  # noqa
+from .group import *  # noqa
+from .image import *  # noqa
+from .kernel import *  # noqa
+from .keypair import *  # noqa
+from .scaling_group import *  # noqa
+from .session import *  # noqa
+from .user import *  # noqa
