+++ src/ai/backend/manager/models/gql_models/agent.py
@@ -49,9 +49,9 @@ from ..base import (
     set_if_set,
     simple_db_mutate,
 )
+from ..gql_models.kernel import ComputeContainer, KernelStatus
 from ..gql_relay import AsyncNode, Connection, ConnectionResolverResult
 from ..group import AssocGroupUserRow
-from ..kernel import ComputeContainer, KernelStatus
 from ..keypair import keypairs
 from ..minilang.ordering import OrderSpecItem, QueryOrderParser
 from ..minilang.queryfilter import FieldSpecItem, QueryFilterParser, enum_field_getter
