+++ src/ai/backend/manager/models/gql_models/domain.py
@@ -77,6 +77,23 @@ if TYPE_CHECKING:
     from .scaling_group import ScalingGroupNode
 
 
+__all__ = (
+    "DomainNode",
+    "DomainConnection",
+    "DomainPermissionValueField",
+    "CreateDomainNodeInput",
+    "CreateDomainNode",
+    "ModifyDomainNodeInput",
+    "ModifyDomainNode",
+    "Domain",
+    "DomainInput",
+    "CreateDomain",
+    "ModifyDomain",
+    "DeleteDomain",
+    "PurgeDomain",
+)
+
+
 class DomainPermissionValueField(graphene.Scalar):
     class Meta:
         description = f"Added in 24.12.0. One of {[val.value for val in DomainPermission]}."
